+++ src/main/java/com/example/demo/service/MemberService.java
@@ -1,9 +1,12 @@
 package com.example.demo.service;
 
 import java.util.List;
+import java.util.Optional;
 
+import com.example.demo.domain.Article;
 import com.example.demo.exception.errorcode.CommonErrorCode;
 import com.example.demo.exception.exception.RestApiException;
+import com.example.demo.repository.ArticleRepository;
 import jakarta.validation.Valid;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -19,9 +22,11 @@ import com.example.demo.repository.MemberRepository;
 public class MemberService {
 
     private final MemberRepository memberRepository;
+    private final ArticleRepository articleRepository;
 
-    public MemberService(MemberRepository memberRepository) {
+    public MemberService(MemberRepository memberRepository, ArticleRepository articleRepository) {
         this.memberRepository = memberRepository;
+        this.articleRepository = articleRepository;
     }
 
     public MemberResponse getById(Long id) {
@@ -52,6 +57,15 @@ public class MemberService {
 
     @Transactional
     public void delete(Long id) {
+        // 삭제 시 article이 참조하고 있는 경우 삭제를 못하게 해야됨.
+        Boolean isArticleExist = articleRepository.findAllByMemberId(id).stream()
+                .findAny()
+                .isPresent();
+
+        if(isArticleExist) {
+            throw new RestApiException(CommonErrorCode.DELETE_ARTICLE_EXIST_IN_MEMBER);
+        }
+
         memberRepository.deleteById(id);
     }
